cudaPOISSON is buggy

+++ Driver_data.F90
@@ -2,8 +2,8 @@ module Driver_data
 
     implicit none
 
-    real, save :: dr_t
-    real, save :: dr_dt
-    integer, save :: dr_nt
+    real*4, save :: dr_t
+    real*4, save :: dr_dt
+    integer*4, save :: dr_nt
 
 end module Driver_data
