+++ MPI_applyBC.F90
@@ -8,10 +8,10 @@ subroutine MPI_applyBC(u_ex)
 
        include "mpif.h"
 
-       real, dimension(Nxb+2,Nyb+2), intent(inout) :: u_ex
-       integer :: status(MPI_STATUS_SIZE)
+       real*4, dimension(Nxb+2,Nyb+2), intent(inout) :: u_ex
+       integer*4 :: status(MPI_STATUS_SIZE)
         
-       integer :: sendcounts(HK), recvcounts(HK), displs(HK)
+       integer*4 :: sendcounts(HK), recvcounts(HK), displs(HK)
 
        sendcounts = 0
        recvcounts = 0
