+++ MPI_physicalBC_vel.F90
@@ -8,8 +8,8 @@ subroutine MPI_physicalBC_vel(u_ex,v_ex)
 
        include "mpif.h"
 
-       real, dimension(Nxb+2,Nyb+2), intent(inout) :: u_ex, v_ex
-       integer :: status(MPI_STATUS_SIZE)
+       real*4, dimension(Nxb+2,Nyb+2), intent(inout) :: u_ex, v_ex
+       integer*4 :: status(MPI_STATUS_SIZE)
 
 
 #ifdef LID_DRIVEN_FLOW
