+++ Multiphase_interface.F90
@@ -13,16 +13,16 @@ implicit none
   interface 
     subroutine Multiphase_solver(tstep,solnX)
      implicit none
-     integer, intent(in) :: tstep
-     real, intent(out) :: solnX
+     integer*4, intent(in) :: tstep
+     real*4, intent(out) :: solnX
     end subroutine Multiphase_solver
   end interface
 
   interface
     subroutine Multiphase_evolve(s,pf,thco,cprs,thco1,thco2,cp1,cp2)
     implicit none
-    real, intent(inout), dimension(Nxb+2,Nyb+2) :: s,pf,thco,cprs
-    real, intent(in) :: thco1,thco2,cp1,cp2
+    real*4, intent(inout), dimension(Nxb+2,Nyb+2) :: s,pf,thco,cprs
+    real*4, intent(in) :: thco1,thco2,cp1,cp2
     end subroutine Multiphase_evolve
   end interface
