+++ Poisson_interface.F90
@@ -7,11 +7,11 @@ module Poisson_interface
        interface
              subroutine Poisson_solver(ps_RHS,ps,ps_res,ps_counter,ps_quant)
                 implicit none
-                real, dimension(Nxb,Nyb), intent(in) :: ps_RHS
-                real, dimension(Nxb+2,Nyb+2), intent(inout) :: ps
-                real, intent(out) :: ps_res
-                integer, intent(out) :: ps_counter
-                integer, intent(in) :: ps_quant
+                real*4, dimension(Nxb,Nyb), intent(in) :: ps_RHS
+                real*4, dimension(Nxb+2,Nyb+2), intent(inout) :: ps
+                real*4, intent(out) :: ps_res
+                integer*4, intent(out) :: ps_counter
+                integer*4, intent(in) :: ps_quant
              end subroutine Poisson_solver
        end interface
